Packet parsers update from Tiv
git-svn-id: https://svn.code.sf.net/p/dolserver/code/PacketLogConverter/trunk@587 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ LogPackets/CtoS_0xF4_CryptKeyRequest.cs
@@ -21,7 +21,7 @@ namespace PacketLogConverter.LogPackets
 
 		#endregion
 
-		public enum eClientType : int
+		public enum eClientType : byte
 		{
 			Classic = 1,
 			ShroudedIsles = 2,
